update : Create new file(
Generates two random numbers less than or equal to MAX.

If the user answers correctly, they continue; if they answer incorrectly, they lose the game.

Even if you lose, the user is asked whether to restart the game and the final decision is made whether to restart or end the game.